[flow] add FileDB.get_table_columns() test
